feat: rename member-ordering-extended to member-ordering
