+++ test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering/examples/alphabetical_example.dart
@@ -10,10 +10,4 @@ class Test {
   void create() {} // LINT
 
   void init() {} // LINT
-
-  @Input()
-  String last;
-
-  @Input() // LINT
-  String first;
 }
