+++ test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering/examples/example.dart
@@ -1,6 +1,14 @@
 class Test {
   String _value;
 
+  static const staticConstField = ''; // LINT
+
+  late final staticLateFinalField; // LINT
+
+  String? nullableField; // LINT
+
+  late String? lateNullableField; // LINT
+
   void doWork() {}
 
   void doAnotherWork() {}
@@ -8,6 +16,10 @@ class Test {
   final data = 1; // LINT
 
   Test();
+
+  factory Test.empty() => Test();
+
+  Test.create();
 
   String _doWork() => 'test';
