+++ test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering/member_ordering_rule_test.dart
@@ -4,15 +4,20 @@ import 'package:test/test.dart';
 
 import '../../../../../helpers/rule_test_helper.dart';
 
-const _examplePath = 'member_ordering/examples/example.dart';
+const _examplePath = 'member_ordering_extended/examples/example.dart';
 const _multipleClassesExamplePath =
-    'member_ordering/examples/multiple_classes_example.dart';
-const _angularExamplePath = 'member_ordering/examples/angular_example.dart';
+    'member_ordering_extended/examples/multiple_classes_example.dart';
 const _alphabeticalExamplePath =
-    'member_ordering/examples/alphabetical_example.dart';
+    'member_ordering_extended/examples/alphabetical_example.dart';
+const _alphabeticalCorrectExamplePath =
+    'member_ordering_extended/examples/alphabetical_correct_example.dart';
+const _typeAlphabeticalExamplePath =
+    'member_ordering_extended/examples/type_alphabetical_example.dart';
+const _flutterMethodsExamplePath =
+    'member_ordering_extended/examples/flutter_methods_example.dart';
 
 void main() {
-  group('MemberOrderingRule', () {
+  group('MemberOrderingExtendedRule', () {
     test('initialization', () async {
       final unit = await RuleTestHelper.resolveFromFile(_examplePath);
       final issues = MemberOrderingRule().check(unit);
@@ -23,6 +28,7 @@ void main() {
         severity: Severity.style,
       );
     });
+
     group('with default config', () {
       test('reports about found issues', () async {
         final unit = await RuleTestHelper.resolveFromFile(_examplePath);
@@ -30,20 +36,28 @@ void main() {
 
         RuleTestHelper.verifyIssues(
           issues: issues,
-          startLines: [8, 16],
-          startColumns: [3, 3],
+          startLines: [4, 6, 8, 10, 16, 28],
+          startColumns: [3, 3, 3, 3, 3, 3],
           locationTexts: [
+            "static const staticConstField = '';",
+            'late final staticLateFinalField;',
+            'String? nullableField;',
+            'late String? lateNullableField;',
             'final data = 1;',
             'String get value => _value;',
           ],
           messages: [
+            'public-fields should be before private-fields.',
+            'public-fields should be before private-fields.',
+            'public-fields should be before private-fields.',
+            'public-fields should be before private-fields.',
             'public-fields should be before public-methods.',
             'public-getters should be before private-methods.',
           ],
         );
       });
 
-      test('and multiple classes reports no issues', () async {
+      test('and multiple classes in a file reports no issues', () async {
         final unit =
             await RuleTestHelper.resolveFromFile(_multipleClassesExamplePath);
         final issues = MemberOrderingRule().check(unit);
@@ -68,30 +82,31 @@ void main() {
 
         RuleTestHelper.verifyIssues(
           issues: issues,
-          startLines: [10, 18],
-          startColumns: [3, 3],
+          startLines: [18, 20, 22, 30],
+          startColumns: [3, 3, 3, 3],
           locationTexts: [
             'Test();',
+            'factory Test.empty() => Test();',
+            'Test.create();',
             'set value(String str) => _value = str;',
           ],
           messages: [
+            'constructors should be before public-fields.',
+            'constructors should be before public-fields.',
             'constructors should be before public-fields.',
             'public-setters should be before private-methods.',
           ],
         );
       });
 
-      test('for angular decorators reports about found issues', () async {
-        final unit = await RuleTestHelper.resolveFromFile(_angularExamplePath);
+      test('accepts categories and reports about found issues', () async {
+        final unit = await RuleTestHelper.resolveFromFile(_examplePath);
         final config = {
           'order': [
-            'angular-outputs',
-            'angular-inputs',
-            'angular-host-listeners',
-            'angular-host-bindings',
-            'angular-content-children',
-            'angular-view-children',
             'constructors',
+            'fields',
+            'getters-setters',
+            'methods',
           ],
         };
 
@@ -99,45 +114,115 @@ void main() {
 
         RuleTestHelper.verifyIssues(
           issues: issues,
-          startLines: [4, 7, 10, 13, 16, 19, 25],
-          startColumns: [3, 3, 3, 3, 3, 3, 3],
+          startLines: [16, 18, 20, 22, 28, 30],
+          startColumns: [3, 3, 3, 3, 3, 3],
           locationTexts: [
-            "@ViewChild('')\n"
-                '  Element view;',
-            "@ViewChild('')\n"
-                '  Iterable<Element> views;',
-            "@ContentChild('')\n"
-                '  Element content;',
-            "@ContentChildren('')\n"
-                '  Iterable<Element> contents;',
-            '@Input()\n'
-                '  String input;',
-            '@Output()\n'
-                '  Stream<void> get click => null;',
-            "@HostListener('')\n"
-                '  void handle() => null;',
+            'final data = 1;',
+            'Test();',
+            'factory Test.empty() => Test();',
+            'Test.create();',
+            'String get value => _value;',
+            'set value(String str) => _value = str;',
           ],
           messages: [
-            'angular-view-children should be before constructors.',
-            'angular-view-children should be before constructors.',
-            'angular-content-children should be before angular-view-children.',
-            'angular-content-children should be before angular-view-children.',
-            'angular-inputs should be before angular-content-children.',
-            'angular-outputs should be before angular-inputs.',
-            'angular-host-listeners should be before angular-host-bindings.',
+            'fields should be before methods.',
+            'constructors should be before fields.',
+            'constructors should be before fields.',
+            'constructors should be before fields.',
+            'getters-setters should be before methods.',
+            'getters-setters should be before methods.',
           ],
         );
       });
 
-      test('for alphabetical order reports about found issues', () async {
+      test(
+        'accepts decomposed constructor categories and reports about found issues',
+        () async {
+          final unit = await RuleTestHelper.resolveFromFile(_examplePath);
+          final config = {
+            'order': [
+              'constructors',
+              'fields',
+              'getters-setters',
+              'named-constructors',
+              'factory-constructors',
+              'methods',
+            ],
+          };
+
+          final issues = MemberOrderingRule(config).check(unit);
+
+          RuleTestHelper.verifyIssues(
+            issues: issues,
+            startLines: [16, 18, 22, 28, 30],
+            startColumns: [3, 3, 3, 3, 3],
+            locationTexts: [
+              'final data = 1;',
+              'Test();',
+              'Test.create();',
+              'String get value => _value;',
+              'set value(String str) => _value = str;',
+            ],
+            messages: [
+              'fields should be before methods.',
+              'constructors should be before fields.',
+              'named-constructors should be before factory-constructors.',
+              'getters-setters should be before methods.',
+              'getters-setters should be before methods.',
+            ],
+          );
+        },
+      );
+
+      test(
+        'accepts decomposed fields categories and reports about found issues',
+        () async {
+          final unit = await RuleTestHelper.resolveFromFile(_examplePath);
+          final config = {
+            'order': [
+              'public-static-fields',
+              'private-static-fields',
+              'public-final-fields',
+              'private-final-fields',
+              'late-fields',
+              'nullable-fields',
+              'private-fields',
+            ],
+          };
+
+          final issues = MemberOrderingRule(config).check(unit);
+
+          RuleTestHelper.verifyIssues(
+            issues: issues,
+            startLines: [4, 10, 16],
+            startColumns: [3, 3, 3],
+            locationTexts: [
+              "static const staticConstField = '';",
+              'late String? lateNullableField;',
+              'final data = 1;',
+            ],
+            messages: [
+              'public-static-fields should be before private-fields.',
+              'late-fields should be before nullable-fields.',
+              'public-final-fields should be before late-fields.',
+            ],
+          );
+        },
+      );
+
+      test('accepts flutter methods and reports issues', () async {
         final unit =
-            await RuleTestHelper.resolveFromFile(_alphabeticalExamplePath);
+            await RuleTestHelper.resolveFromFile(_flutterMethodsExamplePath);
         final config = {
-          'alphabetize': true,
           'order': [
+            'constructors',
+            'overridden-methods',
+            'dispose-method',
+            'did-update-widget-method',
+            'did-change-dependencies-method',
+            'init-state-method',
+            'build-method',
             'public-methods',
-            'public-fields',
-            'angular-inputs',
           ],
         };
 
@@ -145,28 +230,129 @@ void main() {
 
         RuleTestHelper.verifyIssues(
           issues: issues,
-          startLines: [8, 10, 12, 4, 6, 10, 17],
-          startColumns: [3, 3, 3, 3, 3, 3, 3],
+          startLines: [7, 12, 15, 18, 21, 24],
+          startColumns: [3, 3, 3, 3, 3, 3],
           locationTexts: [
-            'void work() {}',
-            'void create() {}',
-            'void init() {}',
-            'final data = 2;',
-            'final algorithm = 3;',
-            'void create() {}',
-            '@Input() // LINT\n'
-                '  String first;',
+            '@override\n'
+                '  Widget build(BuildContext context) {\n'
+                '    return Widget();\n'
+                '  }',
+            '@override\n'
+                '  void initState() {}',
+            '@override\n'
+                '  void didChangeDependencies() {}',
+            '@override\n'
+                '  void didUpdateWidget() {}',
+            '@override\n'
+                '  void dispose() {}',
+            '@override\n'
+                '  void someOtherMethod() {}',
           ],
           messages: [
-            'public-methods should be before public-fields.',
-            'public-methods should be before public-fields.',
-            'public-methods should be before public-fields.',
-            'data should be alphabetically before value.',
-            'algorithm should be alphabetically before data.',
-            'create should be alphabetically before work.',
-            'first should be alphabetically before last.',
+            'build-method should be before public-methods.',
+            'init-state-method should be before build-method.',
+            'did-change-dependencies-method should be before init-state-method.',
+            'did-update-widget-method should be before did-change-dependencies-method.',
+            'dispose-method should be before did-update-widget-method.',
+            'overridden-methods should be before dispose-method.',
           ],
         );
+      });
+
+      group('and alphabetical order', () {
+        test('reports about found issues', () async {
+          final unit =
+              await RuleTestHelper.resolveFromFile(_alphabeticalExamplePath);
+          final config = {
+            'alphabetize': true,
+            'order': [
+              'public-methods',
+              'public-fields',
+            ],
+          };
+
+          final issues = MemberOrderingRule(config).check(unit);
+
+          RuleTestHelper.verifyIssues(
+            issues: issues,
+            startLines: [8, 10, 12, 4, 6, 10],
+            startColumns: [3, 3, 3, 3, 3, 3],
+            locationTexts: [
+              'void work() {}',
+              'void create() {}',
+              'void init() {}',
+              'final data = 2;',
+              'final algorithm = 3;',
+              'void create() {}',
+            ],
+            messages: [
+              'public-methods should be before public-fields.',
+              'public-methods should be before public-fields.',
+              'public-methods should be before public-fields.',
+              'data should be alphabetically before value.',
+              'algorithm should be alphabetically before data.',
+              'create should be alphabetically before work.',
+            ],
+          );
+        });
+
+        test('reports no issues', () async {
+          final unit = await RuleTestHelper.resolveFromFile(
+            _alphabeticalCorrectExamplePath,
+          );
+          final config = {
+            'alphabetize': true,
+            'order': [
+              'public-fields',
+              'public-getters-setters',
+              'private-fields',
+              'private-getters-setters',
+            ],
+          };
+
+          final issues = MemberOrderingRule(config).check(unit);
+
+          RuleTestHelper.verifyNoIssues(issues);
+        });
+      });
+
+      group('and type alphabetical order', () {
+        test('reports about found issues', () async {
+          final unit = await RuleTestHelper.resolveFromFile(
+            _typeAlphabeticalExamplePath,
+          );
+          final config = {
+            'alphabetize-by-type': true,
+            'order': [
+              'public-methods',
+              'public-fields',
+            ],
+          };
+
+          final issues = MemberOrderingRule(config).check(unit);
+
+          RuleTestHelper.verifyIssues(
+            issues: issues,
+            startLines: [15, 19, 21, 8, 19],
+            startColumns: [3, 3, 3, 3, 3],
+            locationTexts: [
+              'String work() {\n'
+                  "    return '';\n"
+                  '  }',
+              'bool create() => false;',
+              'void init() {}',
+              'double f;',
+              'bool create() => false;',
+            ],
+            messages: [
+              'public-methods should be before public-fields.',
+              'public-methods should be before public-fields.',
+              'public-methods should be before public-fields.',
+              'f type name should be alphabetically before z.',
+              'create type name should be alphabetically before work.',
+            ],
+          );
+        });
       });
     });
   });
