--- test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering_extended/examples/alphabetical_example.dart
@@ -1,13 +1,0 @@
-class Test {
-  final value = 1;
-
-  final data = 2; // LINT
-
-  final algorithm = 3; // LINT
-
-  void work() {} // LINT
-
-  void create() {} // LINT
-
-  void init() {} // LINT
-}
