--- test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering_extended/examples/multiple_classes_example.dart
@@ -1,15 +1,0 @@
-class Test {
-  final _data = 1;
-
-  int get data => _data;
-
-  void doWork() {}
-}
-
-class AnotherTest {
-  final _anotherData = 1;
-
-  int get anotherData => _anotherData;
-
-  void anotherDoWork() {}
-}
