+++ website/docs/rules/index.mdx
@@ -251,16 +251,6 @@ Rules are grouped by category to help you understand their purpose. Each rule ha
 </RuleEntry>
 
 <RuleEntry
-  name="member-ordering-extended"
-  type="common"
-  severity="style"
-  version="3.0.0"
-  hasConfig
->
-  Enforces ordering for a class members.
-</RuleEntry>
-
-<RuleEntry
   name="missing-test-assertion"
   type="common"
   severity="warning"
